refactored client subscription publish strategy
